Skiller ut reaktiveringRepository i egen klasse/interface
Hensikten er å gjøre dette tettere koblet til reaktiveringflyten.
