+++ src/main/java/no/nav/fo/veilarbregistrering/db/RepositoryConfig.java
@@ -6,19 +6,25 @@ import no.nav.fo.veilarbregistrering.db.oppgave.OppgaveRepositoryImpl;
 import no.nav.fo.veilarbregistrering.db.profilering.ProfileringRepositoryImpl;
 import no.nav.fo.veilarbregistrering.db.registrering.BrukerRegistreringRepositoryImpl;
 import no.nav.fo.veilarbregistrering.db.registrering.ManuellRegistreringRepositoryImpl;
+import no.nav.fo.veilarbregistrering.db.registrering.ReaktiveringRepositoryImpl;
 import no.nav.fo.veilarbregistrering.db.registrering.RegistreringTilstandRepositoryImpl;
 import no.nav.fo.veilarbregistrering.oppgave.OppgaveRepository;
 import no.nav.fo.veilarbregistrering.profilering.ProfileringRepository;
 import no.nav.fo.veilarbregistrering.registrering.bruker.BrukerRegistreringRepository;
+import no.nav.fo.veilarbregistrering.registrering.bruker.ReaktiveringRepository;
 import no.nav.fo.veilarbregistrering.registrering.manuell.ManuellRegistreringRepository;
 import no.nav.fo.veilarbregistrering.registrering.tilstand.RegistreringTilstandRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
 @Configuration
 public class RepositoryConfig {
+
+    @Bean
+    ReaktiveringRepository reaktiveringRepository(NamedParameterJdbcTemplate db) {
+        return new ReaktiveringRepositoryImpl(db);
+    }
 
     @Bean
     BrukerRegistreringRepository brukerRegistreringRepository(NamedParameterJdbcTemplate db) {
