+++ src/main/java/no/nav/fo/veilarbregistrering/registrering/bruker/ReaktiveringRepository.kt
@@ -1,0 +1,8 @@
+package no.nav.fo.veilarbregistrering.registrering.bruker
+
+import no.nav.fo.veilarbregistrering.bruker.AktorId
+
+interface ReaktiveringRepository {
+
+    fun lagreReaktiveringForBruker(aktorId: AktorId)
+}
